+++ app/models/transfer.rb
@@ -9,5 +9,11 @@ class Transfer < ApplicationRecord
   belongs_to :delivery
   has_many :transfer_products
   accepts_nested_attributes_for :transfer_products, reject_if: :all_blank,allow_destroy: true
-
+  def total_items
+	total_items = 0
+	transfer_products.each do |transfer_product|
+		total_items += transfer_product.quantity
+	end
+	total_items
+  end
 end