+++ db/migrate/20180212125448_create_transfers.rb
@@ -1,8 +1,8 @@
 class CreateTransfers < ActiveRecord::Migration[5.1]
   def change
     create_table :transfers do |t|
-      t.references :store_to
-      t.references :store_from
+      t.references :store_to, null: false
+      t.references :store_from, null: false
       t.references :pullout, foreign_key: true
       t.references :delivery, foreign_key: true